v0.1.2 - Corrected invalid memory usage, that could cause a crash when using   long names for default_theme. - Removed config reloading on SIGHUP because it was not working - Now a border can be drawn around the whole bar. - osd_default_theme in rc.conf can be a real number. - Added nice configuration examples. - Revised build system for simpler version control.

+++ src/config.h
@@ -20,7 +20,7 @@ typedef struct
 
 #define MAX_THEME_LEN 64
 extern char         _default_theme[];
-extern unsigned int _osd_default_timeout;
+extern double       _osd_default_timeout;
 
 #ifdef _GRAPHICAL_
